requirements.txt
python-dotenv>=1.0.0
google-genai>=0.2.0
requests>=2.31.0
